requirements.txt
firebase_admin==5.2.0
Flask
Flask_Cors
requests==2.22.0
python-dotenv==0.20.0